refactor to remove duplicate code in handlers
