+++ src/main/java/com/redhat/devtools/intellij/tektoncd/completion/BaseAutoInsertHandler.java
@@ -10,11 +10,27 @@
  ******************************************************************************/
 package com.redhat.devtools.intellij.tektoncd.completion;
 
+import com.intellij.application.options.CodeStyle;
+import com.intellij.codeInsight.completion.InsertHandler;
+import com.intellij.codeInsight.completion.InsertionContext;
+import com.intellij.codeInsight.lookup.LookupElement;
 import com.intellij.openapi.editor.Document;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Collections;
 
-public class BaseAutoInsertHandler {
+public abstract class BaseAutoInsertHandler implements InsertHandler<LookupElement> {
+    @Override
+    public void handleInsert(@NotNull InsertionContext context, @NotNull LookupElement item) {
+        Document document = context.getEditor().getDocument();
+        int startOffset = context.getStartOffset();
+        int tailOffset = context.getTailOffset();
+        int indentationSize = CodeStyle.getIndentOptions(context.getProject(), document).INDENT_SIZE;
+        int indentationParent = getParentIndentation(document, getParentName(), startOffset);
+        String completionText = getCompletionText(item, indentationSize, indentationParent);
+        document.replaceString(startOffset, tailOffset, completionText);
+    }
+
     protected int getParentIndentation(Document document, String label, int offset) {
         int positionNewLine = document.getText().substring(0, offset).lastIndexOf("\n");
         int nSpaces = document.getText().indexOf(label, positionNewLine);
@@ -24,4 +40,8 @@ public class BaseAutoInsertHandler {
     protected String getIndentationAsText(int indentationParent, int indentSize, int level) {
         return String.join("", Collections.nCopies(indentationParent + (indentSize * level), " "));
     }
+
+    public abstract String getParentName();
+
+    public abstract String getCompletionText(LookupElement item, int indentationSize, int indentationParent);
 }
