+++ src/main/java/com/redhat/devtools/intellij/tektoncd/completion/ConditionAutoInsertHandler.java
@@ -10,26 +10,21 @@
  ******************************************************************************/
 package com.redhat.devtools.intellij.tektoncd.completion;
 
-import com.intellij.application.options.CodeStyle;
-import com.intellij.codeInsight.completion.InsertHandler;
-import com.intellij.codeInsight.completion.InsertionContext;
 import com.intellij.codeInsight.lookup.LookupElement;
-import com.intellij.openapi.editor.Document;
 import io.fabric8.tekton.pipeline.v1alpha1.Condition;
 import io.fabric8.tekton.resource.v1alpha1.ResourceDeclaration;
 import io.fabric8.tekton.v1alpha1.internal.pipeline.pkg.apis.pipeline.v1alpha2.ParamSpec;
-import org.jetbrains.annotations.NotNull;
 
-public class ConditionAutoInsertHandler extends BaseAutoInsertHandler implements InsertHandler<LookupElement> {
+public class ConditionAutoInsertHandler extends BaseAutoInsertHandler {
 
     @Override
-    public void handleInsert(@NotNull InsertionContext context, @NotNull LookupElement item) {
-        Document document = context.getEditor().getDocument();
+    public String getParentName() {
+        return "- conditionRef";
+    }
+
+    @Override
+    public String getCompletionText(LookupElement item, int indentationSize, int indentationParent) {
         Condition condition = (Condition) item.getObject();
-        int startOffset = context.getStartOffset();
-        int tailOffset = context.getTailOffset();
-        int indentationSize = CodeStyle.getIndentOptions(context.getProject(), document).INDENT_SIZE;
-        int indentationParent = getParentIndentation(document, "- conditionRef", startOffset);
 
         String completionText = condition.getMetadata().getName() + "\n";
         if (condition.getSpec().getParams() != null && condition.getSpec().getParams().size() > 0) {
@@ -46,6 +41,7 @@ public class ConditionAutoInsertHandler extends BaseAutoInsertHandler implements
                 completionText += getIndentationAsText(indentationParent, indentationSize, 2) + "resource: \n";
             }
         }
-        document.replaceString(startOffset, tailOffset, completionText);
+
+        return completionText;
     }
 }
